Remove concept of an endpoint from BaseElasticConnection.
