+++ Source/ElasticLINQ/Test/TestableElasticQuery.cs
@@ -62,7 +62,7 @@ namespace ElasticLinq.Test
             var prefix = Context.Mapping.GetDocumentMappingPrefix(typeof(T));
             var request = ElasticQueryTranslator.Translate(Context.Mapping, prefix, Expression);
             var formatter = new SearchRequestFormatter(Context.Connection, Context.Mapping, request.SearchRequest);
-            return new QueryInfo(formatter.Body, formatter.Uri);
+			return new QueryInfo(formatter.Body, Context.Connection.GetSearchUri(request.SearchRequest));
         }
     }
 }
